perf(api): Injector API updated to inject via an 'inject' prop.
The Injector no longer supports the 'elements' prop, it now has an 'inject' prop which allows you to
pass through either a stateless component (that will recieve the same props as the component being
wrapped) or an element.  Using the stateless component method is preferable as it will only be
executed when needed, whereas passing an element requires you to actually create the element up
front.

closes #1 BREAKING CHANGE: The Injector no longer supports the 'elements' prop, it now has an
'inject' prop which allows you to pass through either a stateless component (that will recieve the
same props as the component being wrapped) or an element. Using the stateless component method is
preferable as it will only be executed when needed, whereas passing an element requires you to
acutally create the element up front.